csv chat bot/requirement.txt
gradio
google-generativeai
langchain-google-genai
langchain-experimental
python-dotenv
pandas
langchain